Move part of Kokkos_OpenMP_Instance.cpp into Kokkos_OpenMP.cpp
